% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{draw_samples}
\alias{draw_samples}
\title{Draw samples}
\usage{
draw_samples(delta, n, sigma = 1)
}
\arguments{
\item{delta}{Real difference in means}

\item{n}{Sample size for each group}

\item{sigma}{Standard deviation, same for both groups}
}
\description{
Draws samples of size \code{n} from two normal distributions.
}
\details{
If \code{delta} has length > 1, then a single value of \code{delta} will be drawn uniformly at random.  This simulates a case where there is heterogeneity in effects across (latent) subpopulations, and each study is run on one of these subpopulations.
}
