% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{draw_studies}
\alias{draw_studies}
\title{Do several studies}
\usage{
draw_studies(N, delta, n, seed = NULL)
}
\arguments{
\item{N}{Number of studies}

\item{delta}{Real effect size, or multiple sizes for a mixed/heterogeneous case}

\item{n}{Sample size (per group)}

\item{seed}{Seed for the RNG; if \code{NULL} (the default) no seed will be used}
}
\value{
A \code{tibble::tibble} with one row per study. Column \code{samples} contains the samples for each study.
}
\description{
Do several studies using draw_samples() and t_test().  This simulates doing a meta-analysis of \code{N} studies.
}
