% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{flatten_to_chr}
\alias{flatten_to_chr}
\title{Convenience function to coerce single and mixed deltas to characters}
\usage{
flatten_to_chr(alist)
}
\description{
Convenience function to coerce single and mixed deltas to characters
}
\examples{
flatten_to_chr(list(0, .2, list(0, .7)))
}
