% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{likelihood_ratio}
\alias{likelihood_ratio}
\title{Calculate the log likelihood ratio}
\usage{
likelihood_ratio(h1, h2, test_output, dataf)
}
\arguments{
\item{h1}{List of \code{call} (created by \code{exprs()}) for H1}

\item{h2}{List of \code{call} for H2}

\item{test_output}{List of \code{call} for the test output}

\item{dataf}{Data frame of simulation results, as returned by \code{many_metas()}}
}
\value{
Dataframe with columns
\item{h1, h2, test_output}{Expression for H1, H2, and test output, as strings}
\item{llr} Log (base 10) likelihood ratio L(H1; d) / L(H2; d)
\item{n_false_h1, n_false_h2}{Number of rows post-filtering where \code{test_output} is true, for H1 and H2}
\item{n_true_h1, n_true_h2}{Number of rows post-filtering where \code{test_output} is false, for H1 and H2}
}
\description{
Calculate the log (base 10) likelihood ratio for the sets of hypotheses \code{h1} and \code{h2}.  Note that the interface is slightly different from \code{p_value()}: for the hypothesis and test output, that function takes bare expressions.  This function takes a list of calls, as created by \code{exprs()}.
}
