% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{many_metas}
\alias{many_metas}
\title{Run the simulation: "Do many meta-analyses"}
\usage{
many_metas(NN, N, delta, n)
}
\arguments{
\item{NN}{Number of times to run the simulation for each condition}

\item{N}{Number of studies to conduct in each simulation run (vector)}

\item{delta}{Real effect size (vector)}

\item{n}{Sample size per group for each simulated study (vector)}
}
\value{
A dataframe containing one row per simulation run and the following columns:
\item{meta_idx}{Index of the simulation run within the condition}
\item{N, delta, n}{Simulation conditions.  To pass mixed conditions, construct \code{delta} as a list of lists.}
\item{studies}{Nested samples and t-test results for each study}
\item{young_slope, schsp_slope, qq_slope}{Slopes for the Young, Schweder and Spjøtvoll, and QQ plots}
\item{f_stat, alpha, f_p, f_comp}{F statistic, alpha threshold, p-value, and inference for the F test of linearity of the QQ plot}
\item{aic_linear, aic_quad, aic_comp}{AIC values for the linear and quadratic regression, and inference, for the AIC test of linearity of the QQ plot}
}
\description{
Top-level function for running the simulation
}
\details{
The condition parameters \code{N}, \code{delta}, and \code{n} can be vectors; the function will automatically construct all combinations.  \code{NN} must have length 1.
}
