% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ-plot}
\usage{
qq_plot(studies)
}
\arguments{
\item{studies}{A dataframe of studies to be plotted, as returned by \code{draw_studies()}. Minimally, a dataframe with columns \code{p.uniform} and \code{p.value}.}
}
\value{
A \code{ggplot2} plot
}
\description{
A QQ-plot of the p-values against a uniform distribution.  Equivalent to Young's p-curve with a rescaled x-axis.
}
\details{
\code{p.uniform} is calculated in \code{draw_studies()} as \code{rank/(max(rank)+1)}. This plot includes a "45-degree line," a line with slope 1 and intercept 0.
}
