% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{qq_slope}
\alias{qq_slope}
\title{Slope of a QQ-plot}
\usage{
qq_slope(studies)
}
\arguments{
\item{studies}{A dataframe of studies, as returned by \code{draw_studies()}. Minimally, a dataframe with columns \code{p.uniform} and \code{p.value}.}
}
\value{
The slope, a length-1 numeric
}
\description{
Calculates the slope of the QQ-plot against the uniform distribution.
}
