% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{quo_as_text}
\alias{quo_as_text}
\title{Convert quosure to text}
\usage{
quo_as_text(quo)
}
\arguments{
\item{quo}{A quosure}
}
\value{
A length-one character vector representing the quosure's call
}
\description{
rlang::quo_as_text() is recommended for this, but abbreviates long calls, and as a result some of the quosures we use have identical results.  Instead, we wrap rlang::quo_text(), taking care of the low default width argument.
}
