% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{schsp_curve}
\alias{schsp_curve}
\title{Schweder and Spjøtvoll's p-value plot}
\usage{
schsp_curve(studies)
}
\arguments{
\item{studies}{A dataframe of studies to be plotted, as returned by \code{draw_studies()}. Minimally, a dataframe with columns \code{rank} and \code{p.value}.}
}
\value{
A \code{ggplot2} plot
}
\description{
Schweder and Spjøtvolls "p-value plot."
}
\details{
Schweder and Spjøtvoll's p-value plot plots the descending rank of each study (1 = largest p-value) against 1-p.  It stands in a 1-1 relationship with Young's p-value plot, and so a QQ-plot; but swaps the axes.
}
