% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{schsp_slope}
\alias{schsp_slope}
\title{Slope of Schweder and Spjøtvoll's p-value plot}
\usage{
schsp_slope(studies)
}
\arguments{
\item{studies}{A dataframe of studies, as returned by \code{draw_studies()}. Minimally, a dataframe with columns \code{rank} and \code{p.value}.}
}
\value{
The slope, a length-1 numeric
}
\description{
Calculates the slope of Schweder and Spjøtvoll's p-value plot using a linear regression.
}
