% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{simonsohn_curve}
\alias{simonsohn_curve}
\title{Simonsohn et al.'s p-curve}
\usage{
simonsohn_curve(studies, ...)
}
\arguments{
\item{studies}{A dataframe of studies to be plotted, as returned by \code{draw_studies()}.  Minimally, a dataframe with column \code{p.value}.}

\item{...}{Other aesthetics, passed at the plot level}
}
\value{
A \code{ggplot2} plot
}
\description{
Simonsohn et al.'s p-curve is equivalent to a histogram with binwidth .01 on the subset of p-values below .05
}
\details{
Very very small p-values (below 1e-90) are trimmed to keep the plot from drawing a value at 0.
}
