% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{t_test}
\alias{t_test}
\title{Tidy version of a t-test}
\usage{
t_test(samples)
}
\arguments{
\item{samples}{List of length 2; the two samples}
}
\value{
A one-row \code{tibble::tibble}
}
\description{
Given two samples, do an unpaired, two-sided t-test and return the results in tidy format. This simulates the analysis for a single study.
}
