% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{young_slope}
\alias{young_slope}
\title{Slope of Young's p-curve}
\usage{
young_slope(studies)
}
\arguments{
\item{studies}{A dataframe of studies, as returned by \code{draw_studies()}. Minimally, a dataframe with columns \code{rank} and \code{p.value}.}
}
\value{
The slope, a length-1 numeric
}
\description{
Calculates the slope of Young's p-curve using a linear regression.
}
