% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{many_metas}
\alias{many_metas}
\title{Run the simulation: "Do many meta-analyses"}
\usage{
many_metas(
  NN,
  N,
  delta,
  n,
  alpha = 0.05,
  tost_delta = 0.1,
  gap_threshold = 0.125
)
}
\arguments{
\item{NN}{Number of times to run the simulation for each condition}

\item{N}{Number of studies to conduct in each simulation run (vector)}

\item{delta}{Real effect size (vector)}

\item{n}{Sample size per group for each simulated study (vector)}

\item{alpha}{Threshold for statistical significance, used in various null hypothesis tests}

\item{tost_delta}{Radius for the equivalence bounds for TOST test of slope: slope is substantially different from 1 if it is outside of the interval \verb{[1-delta, 1+delta]}.}

\item{gap_threshold}{Threshold for a "gap" in a p-value plot to count as "gappy"}
}
\value{
A dataframe containing one row per simulation run and the following columns:
\item{meta_idx}{Index of the simulation run within the condition}
\item{N, delta, n}{Simulation conditions.  To pass mixed conditions, construct \code{delta} as a list of lists.}
\item{studies}{Nested samples and t-test results for each study}
\item{young_slope, schsp_slope, qq_slope}{Slopes for the Young, Schweder and Spjøtvoll, and QQ plots}
\item{f_stat, alpha, f_p, f_comp}{F statistic, alpha threshold, p-value, and inference for the F test of linearity of the QQ plot}
\item{aic_linear, aic_quad, aic_comp}{AIC values for the linear and quadratic regression, and inference, for the AIC test of linearity of the QQ plot}
}
\description{
Top-level function for running the simulation
}
\details{
The condition parameters \code{N}, \code{delta}, and \code{n} can be vectors; the function will automatically construct all combinations.  \code{NN} must have length 1.
}
\examples{
## Run 15 simulations with 12 studies per simulation, a real effect size
## of 0.10, and 27 samples per group; total of
## 15*12 = 180 studies and 15*12*27*2 = 9720 samples
many_metas(15, 12, 0.10, 27)

## Run 15 simulations for each of the 2*2*2 = 8 combinations of these
## parameter settings: 10 or 20 studies per simulation; real effect
## size of 0.10 or 0.20; 27 or 52 samples per group
many_metas(15, c(12, 20), c(.10, .20), c(27, 52))

## Run 15 simulations with 20 studies per simulation, 30 samples per group,
## and a mixed or heterogeneous real effect such that half of the
## population has an effect of 0.10 and half of the population has an
## effect of 0.20
## Note that this mixed condition requires passing `delta` as a list of
## lists rather than a flat vector
many_metas(15, 20, list(list(.1, .2)), 30)

}
