% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-curve.R
\name{qq_slope}
\alias{qq_slope}
\title{Slope of a QQ-plot}
\usage{
qq_slope(studies, alpha = 0.05, delta = 0.1)
}
\arguments{
\item{studies}{A dataframe of studies, as returned by \code{draw_studies()}. Minimally, a dataframe with columns \code{p.uniform} and \code{p.value}.}

\item{alpha}{Threshold for statistical significance for the tests; default 0.05}

\item{delta}{Radius for the equivalence bounds for TOST: slope is substantially different from 1 if it is outside of the interval \verb{[1-delta, 1+delta]}. Default is 0.1.}
}
\value{
A one-row dataframe with the columns
\item{slope_estimate, slope_std.error}{Estimated slope and its standard error}
\item{slope_t.statistic, slope_t.p.value}{Difference between \code{slope_estimate} and 1 in standard error units; statistic for t-test; its p-value}
\item{slope_t.comp}{t-test call for whether the slope is stat. sig. different from 1}
\item{slope_tost.t1, slope_tost.p1, slope_tost.t2, slope_tost.p2}{T-statistics and p-values for one-sample TOST}
\item{slope_tost.p.value}{Combined p-value for one-sample TOST}
\item{slope_tost.comp}{TOST call for whether the slope is equivalent to 1}
\item{slope_ks.stat, slope_ks.p}{Statistic and p-value of the KS test}
\item{slope_ks.comp}{Inference from the KS stat; one of \code{non-uniform} or \code{uniform}}
}
\description{
Calculates the slope of the QQ-plot using a linear regression, then compares it against the uniform distribution using a Kolmogorov-Smirnov test, t-test against the null that the slope is exactly 1, and TOST against the null that the slope is substantially different from 1.
}
